requirements.txt
Flask>=0.10.1
gunicorn
autoenv
psycopg2
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
Flask-WTF
Flask-Testing
SQLAlchemy
alembic
simplejson
anyjson
Werkzeug
Mako
flask-marshmallow
marshmallow-sqlalchemy
bcrypt
Flask-Bcrypt
itsdangerous
pyjwt
cffi
pycparser